Improved Entities
Entities now track which chunk they are in.
Sheep now generate with terrain, and move randomly.
Targeting entity data is now slower if all you have is the object, but this should only be necessary if the entity is being permanently removed.
Also fixed camera movement to be relative to player speed.